+++ test/Sentry.Unity.Tests/ScriptableSentryUnityOptionsTests.cs
@@ -44,7 +44,6 @@ public class ScriptableSentryUnityOptionsTests
             Enabled = false,
             Dsn = "test",
             CaptureInEditor = false,
-            EnableLogDebouncing = true,
             TracesSampleRate = 1.0f,
             AutoSessionTracking = false,
             AutoSessionTrackingInterval = TimeSpan.FromSeconds(1),
@@ -70,7 +69,6 @@ public class ScriptableSentryUnityOptionsTests
         scriptableOptions.Enabled = expectedOptions.Enabled;
         scriptableOptions.Dsn = expectedOptions.Dsn;
         scriptableOptions.CaptureInEditor = expectedOptions.CaptureInEditor;
-        scriptableOptions.EnableLogDebouncing = expectedOptions.EnableLogDebouncing;
         scriptableOptions.TracesSampleRate = (double)expectedOptions.TracesSampleRate;
         scriptableOptions.AutoSessionTracking = expectedOptions.AutoSessionTracking;
         scriptableOptions.AutoSessionTrackingInterval = (int)expectedOptions.AutoSessionTrackingInterval.TotalMilliseconds;
@@ -166,7 +164,6 @@ public class ScriptableSentryUnityOptionsTests
         Assert.AreEqual(expected.Enabled, actual.Enabled);
         Assert.AreEqual(expected.Dsn, actual.Dsn);
         Assert.AreEqual(expected.CaptureInEditor, actual.CaptureInEditor);
-        Assert.AreEqual(expected.EnableLogDebouncing, actual.EnableLogDebouncing);
         Assert.AreEqual(expected.TracesSampleRate, actual.TracesSampleRate);
         Assert.AreEqual(expected.AutoSessionTracking, actual.AutoSessionTracking);
         Assert.AreEqual(expected.AutoSessionTrackingInterval, actual.AutoSessionTrackingInterval);